Texture renamed to Texture2D

+++ Editor/Source/Editor/Controls/Image.cpp
@@ -3,40 +3,40 @@
 #include <imgui.h>
 #include <imgui_internal.h>
 
-#include <Core/Content/Texture.h>
+#include <Core/Content/Texture2D.h>
 
 #include "ControlList.h"
 
 namespace Editor
 {
-    Image::Image() : Control() {}
+	Image::Image() : Control() {}
 
-    Image::Image(int width, int height) : Control()
-    {
-        _width = width;
-        _height = height;
-    }
+	Image::Image(int width, int height) : Control()
+	{
+		_width = (float)width;
+		_height = (float)height;
+	}
 
-    Image::~Image() {}
+	Image::~Image() {}
 
-    void Image::update()
-    {
-        if (!_visible) return;
+	void Image::update()
+	{
+		if (!_visible) return;
 
-        float w = _width;
-        float h = _height;
+		float w = _width;
+		float h = _height;
 
-        if (w == 0) w = ImGui::GetContentRegionAvail().x;
-        if (h == 0) h = ImGui::GetContentRegionAvail().y;
+		if (w == 0) w = ImGui::GetContentRegionAvail().x;
+		if (h == 0) h = ImGui::GetContentRegionAvail().y;
 
-        unsigned int texId = nativeTextureId;
-        if (texture != nullptr) texId = texture->getNativeId();
+		unsigned int texId = nativeTextureId;
+		if (texture != nullptr) texId = texture->getNativeId();
 
-        ImGui::Image((ImTextureID)texId, ImVec2(w, h), ImVec2(0, 1), ImVec2(1, 0));
-    }
+		ImGui::Image((ImTextureID)texId, ImVec2(w, h), ImVec2(0, 1), ImVec2(1, 0));
+	}
 
-    int Image::getControlType() const
-    {
-        return CONTROL_IMAGE;
-    }
+	int Image::getControlType() const
+	{
+		return CONTROL_IMAGE;
+	}
 } // namespace Editor
